requirements.txt
scipy==1.11.4
matplotlib==3.8.2
open3d==0.18.0
ftfy==6.1.3
tensorboardx==2.6.2.2
tqdm==4.66.1
imageio==2.33.1
plyfile==1.0.2
opencv-python==4.6.0.66
sharedarray==3.2.4
omegaconf==2.3.0
hydra-core==1.3.2
neptune==1.8.6
tensorflow==2.15.0
# torch==2.1.2
torchmetrics==1.3.0.post0
# torchvision==0.16.2
lightning==2.0.8
lightning-bolts==0.7.0
pytorch-lightning==1.9.5
neptune==1.8.6
seaborn==0.13.2
spacy==3.7.6
spacy-experimental
clip@git+https://github.com/openai/CLIP.git
openai==1.46.0
open_clip_torch==2.24.0
timm>=0.6.7
regex==2023.12.25
nerfstudio>=1.0.3
easydict==1.13
transformers==4.40.1
objgraph==3.6.1
pympler==1.0.1
click==8.1.7
torch-scatter==2.1.2
python-dotenv==1.0.1
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
git+https://github.com/kerrj/lerf